Modularizes backbone
