+++ petal/mining/modules/__init__.py
@@ -1,1 +1,2 @@
 from .wikipedia import WikipediaModule
+from .backbone  import BackboneModule
